pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "quant-assistant"
version = "0.1.0"
description = "Daily quant digest + QR research scaffolding"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "beautifulsoup4>=4.14.2",
    "feedparser>=6.0.12",
    "jinja2>=3.1.6",
    "cachetools>=5.3.3",
    "markdown>=3.6",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.23.0",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "ruff>=0.14.0",
    "resend>=0.0.0",
    "loguru>=0.7.2",
    "pydantic-settings>=2.4.0",
    "aiohttp-retry>=2.8.3",
    "PyYAML>=6.0.1",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "wrds>=3.1.6",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/assistant"]

[tool.ruff]
line-length = 100